Create README.md
This game is a simple cat-and-mouse game. The user can move the cat around the screen with the mouse. If the cat eats the food, the user wins. If the cat collides with the mouse, the user loses.

The game is implemented using the Pygame library. Pygame is a cross-platform library that provides a simple API for creating graphical games.

The game is a simple example of how to use Pygame to create a graphical game. It can be used as a starting point for creating more complex games.